+++ src/pages/UsersPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { UsersForm } from "../common/components/UsersForm/UsersForm";
+import { UsersForm } from "../users/components/UsersForm/UsersForm";
 
 export const UsersPage = () => {
   return <UsersForm />;
