+++ src/users/selectors/userSelector.ts
@@ -1,0 +1,3 @@
+import { State } from "../../State";
+
+export const userSelector = (state: State) => state.users;
